Fix group attachments (upload)

+++ src/main/java/ua/ukma/nc/service/impl/GroupAttachmentServiceImpl.java
@@ -37,7 +37,7 @@ public class GroupAttachmentServiceImpl implements GroupAttachmentService{
     }
 
     @Override
-    public int createGroupAttachment(GroupAttachment groupAttachment) {
+    public Long createGroupAttachment(GroupAttachment groupAttachment) {
         return groupAttachmentDao.createGroupAttachment(groupAttachment);
     }
